+++ apps/docs/src/components/content/design-tokens/index.js
@@ -4,3 +4,4 @@ export * from './DesignTokenContext';
 export * from './designTokenUtils';
 export * from './docsTokenUtils';
 export * from './figmaTokenUtils';
+export * from './grommetThemeHpeUtils';
